+++ src/App.jsx
@@ -2,9 +2,9 @@ import Navigation from './routes/Navigation'
 
 const App = () => {
     return (
-        <div>
+        <>
             <Navigation />
-        </div>
+        </>
     )
 }
