+++ src/components/Login/LoginForm/LoginForm.jsx
@@ -1,9 +1,12 @@
 import { Form } from 'formik'
-import LoginField from './LoginField'
+import LoginField from '../LoginField/LoginField'
+import styles from './LoginForm.module.css'
+
+const { flex, width, mAuto } = styles
 
 const LoginForm = () => {
     return (
-        <Form>
+        <Form className={`${flex} ${width} ${mAuto}`}>
             <LoginField name="firstName" labelName="First Name" />
             <LoginField name="lastName" labelName="Last Name" />
             <LoginField name="email" labelName="Email" />
