+++ src/routes/Navigation.module.css
@@ -1,0 +1,5 @@
+.layout {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+}
